fix(livery-embed): don't enable ineffective fullscreen control
The fallback to the fullscreen API (e.g: used in iOS) to just maximize the element within its window will have no effect when the embed page is shown within an iframe on another page. It is already shown at full size within the iframe. So let's just not enable that control in those cases.

LLS-1903